Fix compilation issue on macOS due to mismatched std::min types (#11717)

+++ compiler/src/iree/compiler/Codegen/Common/TransformDialectStrategiesGPU.cpp
@@ -277,9 +277,9 @@ void ReductionStrategyStagedThreadDistribution::compute(
   // TODO: Split to ensure 4 on most of the problem and use a 1-epilogue.
   int64_t reductionDimensionSize = captures.reductionOpSizes.back();
   if (reductionDimensionSize > 0 && reductionDimensionSize % 4 == 0)
-    vectorSize = std::min(maxVectorSize, 4l);
+    vectorSize = std::min(maxVectorSize, int64_t(4));
   else if (reductionDimensionSize > 0 && reductionDimensionSize % 2 == 0)
-    vectorSize = std::min(maxVectorSize, 2l);
+    vectorSize = std::min(maxVectorSize, int64_t(2));
   else
     vectorSize = 1;
   // Tile reduction to the maximal multiple `vectorSize` allowed.
